third_party/blink/web_tests/external/wpt/wasm/webapi/esm-integration/js-wasm-cycle-errors.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Check cyclic linking between JavaScript and WebAssembly where JavaScript is higher in the module graph.
  assert_equals: expected 10 but got 5
Harness: the test ran to completion.
